Refactored RabbitMQ connection code - now making it so we can use it for both consumers and publishers.
