--- src/interfaces/MessageQueueChannelWrapperInterface.ts
@@ -1,3 +1,0 @@
-export interface MessageQueueChannelWrapperInterface {
-    subscribe(): Promise<boolean>;
-}
