+++ src/interfaces/RabbitMQSubscriberInterface.ts
@@ -1,0 +1,3 @@
+export interface RabbitMQSubscriberInterface {
+    subscribe(): Promise<boolean>;
+}
